v0.0.4
Major changes:

Complete redesign of how towers are placed. Created purchase and cancel button in top left corner of the screen(always visible). The cancel button is only interactable when the player is placing towers. Spawn a "dummy" tower that follows the players mouse until cancel is pressed. After pressing the buy button, when clicking on a placement object, a tower will be created on the placement object.

Enhanced UI - Integrated money and lives into the UI, not functional yet. Made functions for increasing and decreasing values for both money and lives. Handling most events through CanasScripts replacement, Overlay.