v0.0.10
Added wave spawning with "strength" values for enemies and "budgets" for waves to randomly select enemies to spawn. When an enemy is added to the enemies to be spawned, it's strength is subtracted from the budget. This is repeated until no enemies can be afforded, aka, the lowest enemy strength is greater than the remaining budget for the round.

Changed the enemy classes to have all properties in a single layer(sprite, collision, script) rather than two seperate layers. This made accessing the code and referencing the enemies easier.

Did a similar process to the projectile as with the enemy. Merged the two seperate layers into one.