v0.0.13
Added the base for modules. Module can be physically spawned and added to the player inventory. The player inventory is currently not accessible in the game. Modules currently have no functionality. Only one module is currently implemented and spawnable for debugging purposes. Modules which are already present in the inventory are not added twice, rather an integer value is incremented in respect to the appropriate item.